Extract shared eval case loader
